Image_Mod/requirements.txt
numpy==2.3.2
opencv_python==4.11.0.86
Pillow==11.3.0
scipy==1.16.1
